Add cleanup unused stacks
User can run a stack and forget about it.
This stack will handle some limited resources. Fo this reason,
the system should notify author of running stacks about expire date
by email, and provide an opportunity to prolong the life time of the stack.
If the author ignores email, the stack should be destroyed.