Add select list of available version of docker images
The users often want to run the special version of docker image.
They do not want to search the right version by themselves. Add the ability
to load available tags of a docker image from the docker hub registry
and provide them as Select List with Life Search to users.