Add independent apply/destroy operation
The long executing terraform command showd be performed completely
asynchronous, because user want to have ability to run many stacks in
the same time.

Add ability to look status of stack. The stack card show the current
status of the operation and status of stack if there is not a performing
operation now.